Nearly everthing together

+++ jobs/prepareEurogatePackage.sh
@@ -1,6 +1,5 @@
 #!/bin/sh
 #
-#build_srm=true
 goUp() {
 #
 # on linux we could use 'realpath'
@@ -15,7 +14,7 @@ goUp() {
      cd $m
   fi
 }
-printf " Checking JVM ........ "
+printf " Checking JVM .......... "
 which java 1>/dev/null 2>/dev/null
 if [ $? -ne 0 ] ; then
   echo "Failed : Java VM not found on PATH" 
@@ -29,7 +28,7 @@ if [ "$x" -lt 4 ] ; then
 fi
 echo "Ok"
 #
-printf " Checking Cells ...... "
+printf " Checking Cells ........ "
 if [ ! -f "../classes/cells.jar" ] ; then
   echo "Failed : cells.jar not found in ../classes"
   exit 4
@@ -40,7 +39,7 @@ printf " Compiling Eurogate .... "
 #
 rm -rf ../store/objectivity52
 rm -rf ../store/bdb
-rm -rf ../spy/JumpingPigs.java
+rm -rf ../spy
 #
 export CLASSPATH
 CLASSPATH=../classes/cells.jar:../..
@@ -60,10 +59,25 @@ printf " Making eurogate.jar ... "
  )
 echo "Done" 
 #
+printf " Compiling client ...... "
+(
+cd ../capi
+make -f Makefile.linux >/dev/null 2>&1
+if [ $? -ne 0 ] ; then
+   echo " Failed"
+   echo "    Please run the folling commands for deatils "
+   echo "    cd capi"
+   echo "    make -f Makefile.linux"
+   exit 5
+fi
+echo "Done" 
+)
+#
+printf " Coping stuff to dest .. "
 goUp
 mkdir -p dist/eurogate 2>/dev/null
 cd dist/eurogate
-mkdir classes jobs eurogatedocs 2>/dev/null
+mkdir classes jobs eurogatedocs config bin 2>/dev/null
 #
 cd ../..
 #
@@ -72,13 +86,20 @@ if [ ! -d classes ] ; then
   exit 4
 fi
 #
+printf "classes "
 cp classes/cells.jar classes/eurogate.jar dist/eurogate/classes
+printf "jobs "
 cp jobs/eurogate.lib.sh jobs/wrapper2.sh jobs/needFulThings.sh  dist/eurogate/jobs
+printf "config "
 cp config/eurogateSetup config/eurogate.batch  dist/eurogate/config
 cd dist/eurogate/jobs
-ln -s wrapper2.sh eurogate
+[ ! -f eurogate ] && ln -s wrapper2.sh eurogate
 cd ../../..
-pwd
-#( tar cf - docs ) | ( cd eurogatedocs ; tar xf - )
+printf "docs "
+( tar cf - docs ) | ( cd dist/eurogate/eurogatedocs ; tar xf - )
+printf "bin "
+cp capi/test capi/eucp dist/eurogate/bin
+( cd dist/eurogate/bin ; [ ! -f eurm ] && ln -s eucp eurm )
 #
+echo "Done"
 exit 0
